RPM Framework Refactoring
This is the beginning of the refactoring of the RMP framework, which
will be entirely re-written to handle mostly back-end functionality. The
front-end will now use Angular.js to handle some of the functions that
the previous versions did server-side.

This first commit contains a reworking of the fabric and database
configuration tools that were used to generate the database for an
application. This is the starting point for the new project, and it has
not been fully tested yet.