Fab To Handle Routes in Server File
This commit marks the beginning of changes that are to be made to the
server portion of fabric that will set up routes for an app based on the
specification in apps.py. This commit contains an example of how they
may look and some changes to the server module that indicate how the
changes are to be implemented.