First Version of Order Request Handler
This commit has the first basic version of the request handler for
orders. We're currently working off of a single request handler to deal
with three types of requests (for post requests at least). As of this
commit, the functionality isn't implemented, but the basic arcitecture
is spelled out here. As well, the route has been added under apps.