Adds Rudimentary Stripe Processing
This adds the beginning of what will be our payment processing through
Stripe. It adds a module that defines a function that posts payments to
Stripe and returns a repsonse from it. As well, it works the code into
the order, but not in a really significant way yet that checks to make
sure that the order has been processed. We'll have to add in something
that checks the output before finalizing the order as the next step in
this portion.