Basic Validation for Stripe Processing
This adds a function to the stripe module that makes it possible to
check the response from stripe when processing an order. Since stripe
returns a dictionary, this function takes that dictionary and a list of
kwargs that specify keys and the values that they should have in order
to validate the payment.

Right now the function checks to make sure that the order has been paid,
not refunded, and not disputed. It's possible to easily add more if
needed.