Server Directory Moved Out of Fabfile
This (finally) moves the server package out of the fabfile package and
into the top level of the project. We had placed the server code in the
fabfile directory earlier on while putting together some rudimentary
server functionality using fabric. It was intended to be moved at some
point--and this is that point.