Moves Server from Fabfile to Server Module
Somewhere along the way in the process of branching, things got messed
up and the server module is still inside the fab file where we had it
while writing and testing it originally. Looking for where the change
occured and what happened is a hassle, so I'm just re-doing that commit
here to incorporate Diego's changes since the re-structure that
apparently never took place.