Removes Path Additions From Tests
This removes the code that was being used to append the path to the
request handlers in this project to the import path for the test scripts
for the request handlers. This is no longer necessary with the addition
of the __init__ file to the top level that extends the path within the
project so that files in adjacent directories are on the import path.